niffler-e-2-e-test-python/requirements.txt
pytest
requests
python-dotenv
selene>=2.0.0rc7
selenium==4.17.2
jsonschema
pydantic
playwright
faker
